+++ person.rb
@@ -1,0 +1,22 @@
+class Person
+  attr_reader :id
+
+  attr_accessor :age, :name
+
+  def initialize(id, age, name = 'unknown', parent_permission: true)
+    @id = id
+    @name = name
+    @age = age
+    @parent_permission = parent_permission
+  end
+
+  def can_use_services?
+    of_age? || @parent_permission
+  end
+
+  private
+
+  def of_age?
+    @age >= 18
+  end
+end
