+++ student.rb
@@ -1,0 +1,13 @@
+require_relative 'person'
+class Student < Person
+  attr_accessor :classroom
+
+  def initialize(id, age, classroom, name = 'unknown', parent_permission: true)
+    super(id, age, name, parent_permission)
+    @class_room = classroom
+  end
+
+  def play_hooky
+    '¯(ツ)/¯'
+  end
+end
